requirements.txt
astroid==2.0.4
blessings==1.7
bpython==0.17.1
certifi==2018.10.15
chardet==3.0.4
curtsies==0.3.0
dj-database-url==0.5.0
Django==2.1.2
django-bootstrap-form==3.4
djangorestframework==3.8.2
greenlet==0.4.15
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.5
Pygments==2.2.0
pylint==2.1.1
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
requests==2.20.0
six==1.11.0
typed-ast==1.1.0
typing==3.6.6
urllib3==1.24
wcwidth==0.1.7
wrapt==1.10.11
